Raise error for non-existent/invalid input images
This should make the script less fragile as far as invalid input is
concerned. Initially I wanted to check the extension against a set of
supported values, but it's easier to just let OpenCV decide whether it
likes the file or not.
Probably not the best approach for many files though and it can be
argued that ignoring unsupported files would be better in some cases.