Split validity check and image reading
Like mentioned before it isn't the best idea to make a hard exit once
OpenCV fails to read an invalid input. That could lead to 999 images
being process before an accidentally specified directory or unsupported
image type throws an error.
So let's check the extension (besides existence and being a file)
beforehand. Not as failsafe as discerning the image type via the file
itself, but good enough for now.