Allow image loading with multiple processes
Now it's possible to read images and calculate their histograms with
multiple processes. This can speed things up considerably on modern
machines (10 sec vs. 1 min with "-t 16" on my Ryzen 2700x for 1000
images). The drawback is the increased RAM usage. 100MB per process are
not uncommon. That's why the script still defaults to a single process.

The command line option -t/--threads is probably a bit misleading, since
we're using several processes and not threads, but I'm used to this
naming convention and I'm not going to change it, just because of
Python's special relationship with multithreading.
Also I can't use -p as short option for --processes, because it's
already the short option for --paths.