Adept sorting algorithm based on comparison method
OpenCV offers several methods to compare the similarity of two given
histograms. The problem is that some of them give higher scores for more
similar histograms (Correlation, Intersection), while others do the
opposite and give a smaller score (Chi-square, Bhattacharyya distance).
Since the script uses Intersection by default, it was only possible to
switch to Correlation without having to change some of the code in the
main function.
The sorting algorithm was moved into a separate function and now decides
whether a min. or max. score is desired based on the used comparison
method, thus allowing the usage of different methods without further
code changes.